refactor: remove createCollapsibleTabs
