+++ example/.eslintrc.json
@@ -1,5 +1,8 @@
 {
   "extends": "../.eslintrc.js",
+  "rules": {
+    "import/no-default-export": "off"
+  },
   "settings": {
     "import/core-modules": [
       "react-native-collapsible-tab-view"
