+++ example/src/OnIndexChange.tsx
@@ -1,16 +1,18 @@
 import React, { useContext } from 'react'
+import {
+  OnTabChangeCallback,
+  TabBarProps,
+} from 'react-native-collapsible-tab-view'
 
-import { OnTabChangeCallback, TabBarProps } from '../../src/types'
 import ExampleComponent from './Shared/ExampleComponent'
 import { Header } from './Shared/Header'
-import { TabNamesType } from './Shared/Tabs'
 import { ExampleComponentType } from './types'
 
 const exampleTitle = 'On index change example'
 
 const TitleContext = React.createContext<string>(exampleTitle)
 
-const HeaderComponent = (props: TabBarProps<TabNamesType>) => {
+const HeaderComponent = (props: TabBarProps) => {
   const title = useContext(TitleContext)
   return <Header title={title} {...props} />
 }
@@ -18,7 +20,7 @@ const HeaderComponent = (props: TabBarProps<TabNamesType>) => {
 const OnIndexChange: ExampleComponentType = () => {
   const [title, setTitle] = React.useState(exampleTitle)
 
-  const onIndexChange = React.useCallback<OnTabChangeCallback<TabNamesType>>(
+  const onIndexChange = React.useCallback<OnTabChangeCallback>(
     ({ prevIndex, index, prevTabName, tabName }) => {
       const title = `prev: ${prevTabName}\ncurr: ${tabName}`
       setTitle(title)
