+++ example/src/ScrollOnHeader.tsx
@@ -1,16 +1,16 @@
 import React from 'react'
 import { View } from 'react-native'
 
+import { TabBarProps } from '../../src/types'
 import ExampleComponent from './Shared/ExampleComponent'
 import Header from './Shared/Header'
-import { HeaderProps } from './Shared/Tabs'
 import { ExampleComponentType } from './types'
 
 const title = 'Scroll On Header'
 
 const HEADER_HEIGHT = 400
 
-const NewHeader: React.FC<HeaderProps> = (props) => {
+const NewHeader: React.FC<TabBarProps> = (props) => {
   return (
     <View pointerEvents="none">
       <Header height={HEADER_HEIGHT} title={title} {...props} />
