+++ example/src/Shared/ExampleComponent.tsx
@@ -1,20 +1,16 @@
 import React from 'react'
-import {
-  CollapsibleProps,
-  CollapsibleRef,
-} from 'react-native-collapsible-tab-view'
+import * as Tabs from 'react-native-collapsible-tab-view'
 
 import Albums from './Albums'
 import Article from './Article'
 import Contacts from './Contacts'
 import { HEADER_HEIGHT } from './Header'
-import Tabs, { TabNamesType } from './Tabs'
 
 type Props = {
   emptyContacts?: boolean
-} & Partial<CollapsibleProps<TabNamesType>>
+} & Partial<Tabs.CollapsibleProps>
 
-const Example = React.forwardRef<CollapsibleRef<TabNamesType>, Props>(
+const Example = React.forwardRef<Tabs.CollapsibleRef, Props>(
   ({ emptyContacts, ...props }, ref) => {
     return (
       <Tabs.Container ref={ref} headerHeight={HEADER_HEIGHT} {...props}>
