+++ example/src/Shared/Header.tsx
@@ -2,8 +2,6 @@ import React from 'react'
 import { StyleSheet, View, Text } from 'react-native'
 import { TabBarProps } from 'react-native-collapsible-tab-view'
 
-import { TabNamesType } from './Tabs'
-
 type Props = {
   title: string
   height?: number
@@ -14,7 +12,7 @@ export const HEADER_HEIGHT = 250
 export const Header = ({
   title,
   height = HEADER_HEIGHT,
-}: TabBarProps<TabNamesType> & Props) => {
+}: TabBarProps & Props) => {
   return (
     <View style={[styles.root, { height }]}>
       <Text style={styles.text}>{title}</Text>
