+++ src/MaterialTabBar/index.tsx
@@ -1,9 +1,3 @@
-import TabBar, { TABBAR_HEIGHT } from './TabBar'
-import TabItem from './TabItem'
-import { MaterialTabBarProps, MaterialTabItemProps } from './types'
-
-export default TabBar
-
-export type { MaterialTabBarProps, MaterialTabItemProps }
-
-export { TabItem as MaterialTabItem, TabBar as MaterialTabBar, TABBAR_HEIGHT }
+export { MaterialTabBar, TABBAR_HEIGHT } from './TabBar'
+export { MaterialTabItem } from './TabItem'
+export type { MaterialTabBarProps, MaterialTabItemProps } from './types'
